requirements.txt
appdirs==1.4.4
argon2-cffi @ file:///C:/ci/argon2-cffi_1613038019788/work
async-generator==1.10
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bleach @ file:///tmp/build/80754af9/bleach_1612211392645/work
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1613247308275/work
click==7.1.2
collection==0.1.6
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
decorator @ file:///tmp/build/80754af9/decorator_1617916966915/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
distlib==0.3.1
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
importlib-metadata @ file:///C:/ci/importlib-metadata_1617877486026/work
ipykernel @ file:///C:/ci/ipykernel_1596208728219/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1617121109687/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
itsdangerous==1.1.0
jedi==0.17.0
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1616770841739/work
jupyter-core @ file:///C:/ci/jupyter_core_1612213516947/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
MarkupSafe @ file:///C:/ci/markupsafe_1594405949945/work
mistune @ file:///C:/ci/mistune_1594373272338/work
mysql==0.0.2
mysqlclient==2.0.3
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///C:/ci/nbconvert_1601914921407/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
notebook @ file:///C:/ci/notebook_1616443616158/work
numpy==1.20.2
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandas==1.2.3
pandocfilters @ file:///C:/ci/pandocfilters_1605102427207/work
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
pipenv==2020.11.15
prettytable==0.7.2
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1617383679942/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1616415428029/work
protobuf==3.15.8
pybars3==0.9.7
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1615143339740/work
PyMeta3==0.5.1
PyMySQL==1.0.2
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyrsistent @ file:///C:/ci/pyrsistent_1600123688363/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
python-dotenv==0.17.0
pytz==2021.1
pywin32==227
pywinpty==0.5.7
pyzmq==20.0.0
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
six @ file:///C:/ci/six_1605205426665/work
SQLAlchemy==1.4.7
terminado==0.9.4
testpath @ file:///home/ktietz/src/ci/testpath_1611930608132/work
tornado @ file:///C:/ci/tornado_1606935947090/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
virtualenv==20.4.3
virtualenv-clone==0.5.4
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
wincertstore==0.2
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
